requirements.txt
kagglehub
torch
numpy
pytorch-minimize
box2d
pygame
gym
